chore(extract): remove old datepicker implementation
