--- libs/extract/src/lib/datepicker/datepicker.ts
@@ -1,465 +1,0 @@
-import {
-  add,
-  addYears,
-  endOfDay,
-  endOfMonth,
-  endOfYear,
-  format,
-  startOfDay,
-  startOfMonth,
-  startOfYear,
-  sub,
-  subYears,
-} from 'date-fns'
-import { Calendar, createCalendar } from '.'
-import { Instance } from '@popperjs/core'
-import { merge, fromEvent, iif, Observable, ReplaySubject, Subject } from 'rxjs'
-import { switchMap, takeUntil } from 'rxjs/operators'
-import { onActiveHandler, onInactiveHandler } from './event-handlers'
-import { setFocus } from './helper-functions'
-import IMask from 'imask'
-
-type DateUnit = 'years' | 'months' | 'weeks' | 'days'
-export interface Datepicker {
-  add: (amount: number, unit: DateUnit, select?: boolean) => void
-  sub: (amount: number, unit: DateUnit, select?: boolean) => void
-  setMonth: (number: number) => void
-  setYear: (number: number) => void
-  set: (date: Date | string) => void
-  highlight: (date: Date | string) => void
-  select: (date: Date | string) => void
-  _popper?: Instance
-  state: DatepickerState
-  open: () => void
-  close: (selectHighlighted?: boolean) => void
-  toggle: () => void
-  destroy: () => void
-  active$: Observable<boolean>
-  selectedDate$: Observable<Date>
-}
-
-export interface DatepickerData {
-  date: Date
-  formattedDate: string
-  year: number
-  month: number
-  monthName: string
-  day: number
-  weekday: string
-  calendar: Calendar
-  selectedDate?: Date
-  formattedSelectedDate: string
-  highlightedDate?: Date
-}
-
-export interface DatepickerState {
-  isActive: boolean
-  minDate?: Date
-  maxDate?: Date
-}
-
-export type DatepickerListener = (
-  data?: DatepickerData,
-  state?: DatepickerState
-) => void
-
-export interface DatepickerOptions {
-  label?: string
-  locale?: string
-  currentDate?: Date | string
-  selectedDate?: Date | string
-  closeOnSelect?: boolean
-  useCurrentTime?: boolean
-  weekName?: { abbr: string; displayText: string }
-  showWeeks?: boolean
-  minDate?: Date
-  maxDate?: Date
-  onChange?: (selectedDate: Date) => void
-}
-
-const createState = (
-  isActive = false,
-  minDate?: Date,
-  maxDate?: Date
-): DatepickerState => {
-  return {
-    isActive,
-    minDate,
-    maxDate,
-  }
-}
-
-const createData = (
-  locale: string,
-  currentDate: Date | string,
-  preSelectedDate?: Date | string,
-  useCurrentTime?: boolean,
-  showWeeks?: boolean,
-  weekName?: { abbr: string; displayText: string },
-  preHighlightedDate?: Date | string | null,
-  startOfMinDate?: Date | undefined,
-  endOfMaxDate?: Date | undefined
-): DatepickerData => {
-  const date =
-    typeof currentDate === 'string'
-      ? new Date(currentDate + ' 12:00:00')
-      : currentDate
-  const formattedDate =
-    typeof currentDate === 'string' ? currentDate : format(date, 'yyyy-MM-dd')
-
-  let selectedDate
-  if (typeof preSelectedDate === 'string') {
-    const preSelectedDateFromString = new Date(preSelectedDate + ' 12:00:00')
-    selectedDate = isNaN(preSelectedDateFromString.getTime())
-      ? undefined
-      : preSelectedDateFromString
-  } else {
-    selectedDate = preSelectedDate
-  }
-  const formattedSelectedDate = preSelectedDate
-    ? typeof preSelectedDate === 'string'
-      ? preSelectedDate
-      : format(preSelectedDate, 'yyyy-MM-dd')
-    : ''
-  const highlightedDate = preHighlightedDate
-    ? typeof preHighlightedDate === 'string'
-      ? new Date(preHighlightedDate + ' 12:00:00')
-      : preHighlightedDate
-    : undefined
-
-  return {
-    selectedDate,
-    formattedSelectedDate,
-    date,
-    formattedDate,
-    highlightedDate,
-    year: date.getFullYear(),
-    month: date.getMonth(),
-    monthName: Intl.DateTimeFormat(locale, { month: 'long' }).format(date),
-    day: date.getDate(),
-    weekday: Intl.DateTimeFormat(locale, { weekday: 'long' }).format(date),
-    calendar: createCalendar(
-      locale,
-      date,
-      selectedDate,
-      <boolean>showWeeks,
-      <boolean>useCurrentTime,
-      <{ abbr: string; displayText: string }>weekName,
-      highlightedDate,
-      startOfMinDate,
-      endOfMaxDate
-    ),
-  }
-}
-
-export const createDatepicker = (
-  listener: DatepickerListener,
-  {
-    selectedDate,
-    locale = navigator.language,
-    currentDate = new Date(),
-    closeOnSelect = true,
-    useCurrentTime = true,
-    showWeeks = false,
-    weekName = { abbr: 'Week', displayText: 'wk' },
-    minDate = startOfYear(subYears(new Date(), 5)),
-    maxDate = endOfYear(addYears(new Date(), 5)),
-  }: DatepickerOptions,
-  datepickerElRef: HTMLElement,
-  datepickerDialogElRef: HTMLElement,
-  // TODO: update value for date input
-  dateInputElRef: HTMLElement,
-  datepickerTriggerElRef: HTMLButtonElement
-): Datepicker => {
-  const startOfMinDate = startOfDay(minDate)
-  const startOfMonthMinDate = startOfMonth(minDate)
-  const endOfMaxDate = endOfDay(maxDate)
-  const endOfMonthMaxDate = endOfMonth(maxDate)
-  let data = createData(
-    locale,
-    currentDate,
-    selectedDate,
-    useCurrentTime,
-    showWeeks,
-    weekName,
-    null,
-    startOfMinDate,
-    endOfMaxDate
-  )
-
-  const unsubscribe$ = new Subject()
-  const active$ = new ReplaySubject<boolean>(1)
-  const selectedDate$ = new Subject<Date>()
-
-  // add input mask to date input
-  // TODO: add support for other date formats
-  IMask(dateInputElRef, {
-    mask: Date,
-    pattern: 'Y{-}`m{-}`z`d',
-    autofix: 'pad',
-    blocks: {
-      d: {
-        mask: IMask.MaskedRange,
-        from: 1,
-        to: 31,
-        maxLength: 2,
-        autofix: false,
-      },
-      z: {
-        mask: IMask.MaskedDynamic,
-        maskFunction: (v: string) => (v === '0' ? '0' : ''),
-      },
-    },
-  })
-
-  const dp: Datepicker = {
-    add: (amount, unit, select = false) => {
-      const newDate = add(data.date, { [unit]: amount })
-      if (unit === 'days' && newDate > endOfMaxDate) {
-        return
-      }
-      if (unit === 'months' && newDate > endOfMonthMaxDate) {
-        return
-      }
-      data = createData(
-        locale,
-        newDate,
-        data.selectedDate,
-        useCurrentTime,
-        showWeeks,
-        weekName,
-        data.highlightedDate,
-        startOfMinDate,
-        endOfMaxDate
-      )
-      if (select) {
-        dp.highlight(data.date)
-      }
-      listener(data)
-    },
-    sub: (amount, unit, select = false) => {
-      const newDate = sub(data.date, { [unit]: amount })
-      if (unit === 'days' && newDate < startOfMinDate) {
-        return
-      }
-      if (unit === 'months' && newDate < startOfMonthMinDate) {
-        return
-      }
-      data = createData(
-        locale,
-        newDate,
-        data.selectedDate,
-        useCurrentTime,
-        showWeeks,
-        weekName,
-        data.highlightedDate,
-        startOfMinDate,
-        endOfMaxDate
-      )
-      if (select) {
-        dp.highlight(data.date)
-      }
-      listener(data)
-    },
-    set: (date) => {
-      if (date < startOfMinDate || date > endOfMaxDate) {
-        return
-      }
-      data = createData(
-        locale,
-        date,
-        data.selectedDate,
-        useCurrentTime,
-        showWeeks,
-        weekName,
-        data.highlightedDate,
-        startOfMinDate,
-        endOfMaxDate
-      )
-      listener(data)
-    },
-    setMonth: (number) => {
-      if (typeof number !== 'number' || number < 0 || number > 11) {
-        return
-      }
-      const date = new Date(data.date.setMonth(number))
-      if (date < startOfMonthMinDate || date > endOfMonthMaxDate) {
-        return
-      }
-      data = createData(
-        locale,
-        date,
-        data.selectedDate,
-        useCurrentTime,
-        showWeeks,
-        weekName,
-        data.highlightedDate,
-        startOfMinDate,
-        endOfMaxDate
-      )
-      listener(data)
-    },
-    setYear: (number) => {
-      if (typeof number !== 'number') {
-        return
-      }
-      const date = new Date(data.date.setFullYear(number))
-      if (date < startOfMonthMinDate || date > endOfMonthMaxDate) {
-        return
-      }
-      data = createData(
-        locale,
-        date,
-        data.selectedDate,
-        useCurrentTime,
-        showWeeks,
-        weekName,
-        data.highlightedDate,
-        startOfMinDate,
-        endOfMaxDate
-      )
-      listener(data)
-    },
-    select: (date) => {
-      if (date < startOfMinDate || date > endOfMaxDate) {
-        return
-      }
-      data = createData(
-        locale,
-        data.date,
-        date,
-        useCurrentTime,
-        showWeeks,
-        weekName,
-        data.highlightedDate,
-        startOfMinDate,
-        endOfMaxDate
-      )
-      listener(data)
-      selectedDate$.next(new Date(date))
-      if (closeOnSelect) {
-        dp.close()
-      }
-    },
-    highlight: (date) => {
-      if (date < startOfMinDate || date > endOfMaxDate) {
-        return
-      }
-      data = createData(
-        locale,
-        data.date,
-        data.selectedDate,
-        useCurrentTime,
-        showWeeks,
-        weekName,
-        date,
-        startOfMinDate,
-        endOfMaxDate
-      )
-      listener(data)
-      // set focus on focusable calendar day
-      setFocus(datepickerDialogElRef, 'td[tabindex="0"]')
-    },
-    open: () => {
-      data = createData(
-        locale,
-        data.selectedDate || new Date(),
-        data.selectedDate,
-        useCurrentTime,
-        showWeeks,
-        weekName,
-        null,
-        startOfMinDate,
-        endOfMaxDate
-      )
-
-      const state = createState(true, startOfMinDate, endOfMaxDate)
-      active$.next(true)
-      listener(data, state)
-
-      // remove focus from trigger element
-      datepickerTriggerElRef.blur()
-
-      // set focus on focusable calendar day
-      setFocus(datepickerDialogElRef, 'td[tabindex="0"]')
-    },
-    close: (selectHighlighted = false) => {
-      if (selectHighlighted) {
-        data = createData(
-          locale,
-          data.date,
-          data.highlightedDate || data.selectedDate,
-          useCurrentTime,
-          showWeeks,
-          weekName,
-          null,
-          startOfMinDate,
-          endOfMaxDate
-        )
-      }
-
-      const state = createState(false, startOfMinDate, endOfMaxDate)
-      active$.next(false)
-      listener(data, state)
-
-      // set focus on trigger
-      datepickerTriggerElRef.focus()
-    },
-    toggle: () => {
-      if (dp.state.isActive) {
-        dp.close()
-      } else {
-        dp.open()
-      }
-    },
-    destroy: () => {
-      unsubscribe$.next(true)
-      unsubscribe$.complete()
-    },
-    state: createState(false, startOfMinDate, endOfMaxDate),
-    active$: active$.asObservable(),
-    selectedDate$: selectedDate$.asObservable(),
-  }
-
-  listener(data, dp.state)
-  dp.active$
-    .pipe(
-      switchMap((active) =>
-        iif(
-          () => active,
-          onActiveHandler(dp, datepickerElRef, datepickerDialogElRef),
-          onInactiveHandler(dp, datepickerDialogElRef)
-        )
-      ),
-      // take until destroyed
-      takeUntil(unsubscribe$)
-    )
-    .subscribe()
-
-  merge(
-    fromEvent<KeyboardEvent>(document, 'keydown'),
-    fromEvent<FocusEvent>(document, 'focusin'),
-    fromEvent<MouseEvent>(document, 'click')
-  )
-    .pipe(takeUntil(unsubscribe$))
-    .subscribe((event) => {
-      if (!dp.state.isActive) return
-
-      switch (event.type) {
-        case 'click':
-        case 'focusin': {
-          const target = event.target as HTMLElement
-          const targetIsOutside =
-            !datepickerElRef.contains(target) &&
-            !datepickerDialogElRef.contains(target) &&
-            !dateInputElRef.contains(target) &&
-            !datepickerTriggerElRef.contains(target)
-          if (targetIsOutside) {
-            dp.close()
-          }
-        }
-      }
-    })
-
-  return dp
-}
