--- libs/extract/src/lib/datepicker/index.ts
@@ -1,3 +1,0 @@
-export * from './calendar'
-export * from './datepicker'
-export { months, years, weekdays, getWeekday } from './values'
